+++ src/Symfony/Component/DependencyInjection/Tests/Config/AutowireServiceResourceTest.php
@@ -14,6 +14,9 @@ namespace Symfony\Component\DependencyInjection\Tests\Config;
 use Symfony\Component\DependencyInjection\Compiler\AutowirePass;
 use Symfony\Component\DependencyInjection\Config\AutowireServiceResource;
 
+/**
+ * @group legacy
+ */
 class AutowireServiceResourceTest extends \PHPUnit_Framework_TestCase
 {
     /**